% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_statistics_for_specific_country}
\alias{get_statistics_for_specific_country}
\title{Download cases history for specific country}
\usage{
get_statistics_for_specific_country(country)
}
\arguments{
\item{country}{string e.g. "Poland", "Italy", etc.}
}
\description{
Download cases history for specific country
}
